+++ e2e/e2etests/test_sui_token_withdraw_and_call_revert_with_call.go
@@ -1,0 +1,82 @@
+package e2etests
+
+import (
+	"math/big"
+
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	"github.com/stretchr/testify/require"
+	"github.com/zeta-chain/protocol-contracts/pkg/gatewayzevm.sol"
+
+	"github.com/zeta-chain/node/e2e/runner"
+	"github.com/zeta-chain/node/e2e/utils"
+	"github.com/zeta-chain/node/testutil/sample"
+	crosschaintypes "github.com/zeta-chain/node/x/crosschain/types"
+)
+
+// TestSuiTokenWithdrawAndCallRevertWithCall executes withdrawAndCall on zevm gateway with fungible token.
+// The outbound is rejected by the connected module due to invalid payload (invalid address),
+// and 'onRevert' is called instead to handle the revert.
+func TestSuiTokenWithdrawAndCallRevertWithCall(r *runner.E2ERunner, args []string) {
+	require.Len(r, args, 1)
+
+	// ARRANGE
+	// Given target package ID (example package) and a token amount
+	targetPackageID := r.SuiExample.PackageID.String()
+	amount := utils.ParseBigInt(r, args[0])
+
+	// create the payload for 'on_call' with invalid address
+	// taking the first 10 letters to form an invalid address
+	invalidAddress := sample.SuiAddress(r)[:10]
+	payloadOnCall, err := r.SuiCreateExampleWACPayload(invalidAddress)
+	require.NoError(r, err)
+
+	// given ZEVM revert address (the dApp)
+	dAppAddress := r.TestDAppV2ZEVMAddr
+	dAppBalanceBefore, err := r.SuiTokenZRC20.BalanceOf(&bind.CallOpts{}, dAppAddress)
+	require.NoError(r, err)
+
+	// given random payload for 'onRevert'
+	payloadOnRevert := randomPayload(r)
+	r.AssertTestDAppEVMCalled(false, payloadOnRevert, amount)
+
+	// ACT
+	// approve both SUI gas budget token and fungible token ZRC20
+	r.ApproveSUIZRC20(r.GatewayZEVMAddr)
+	r.ApproveFungibleTokenZRC20(r.GatewayZEVMAddr)
+
+	// perform the withdraw and call with revert options
+	tx := r.SuiWithdrawAndCallFungibleToken(
+		targetPackageID,
+		amount,
+		payloadOnCall,
+		gatewayzevm.RevertOptions{
+			CallOnRevert:     true,
+			RevertAddress:    dAppAddress,
+			RevertMessage:    []byte(payloadOnRevert),
+			OnRevertGasLimit: big.NewInt(0),
+		},
+	)
+	r.Logger.EVMTransaction(*tx, "withdraw_and_call")
+
+	// ASSERT
+	// wait for the cctx to be reverted
+	cctx := utils.WaitCctxMinedByInboundHash(r.Ctx, tx.Hash().Hex(), r.CctxClient, r.Logger, r.CctxTimeout)
+	r.Logger.CCTX(*cctx, "withdraw")
+	utils.RequireCCTXStatus(r, cctx, crosschaintypes.CctxStatus_Reverted)
+
+	// should have called 'onRevert'
+	r.AssertTestDAppZEVMCalled(true, payloadOnRevert, big.NewInt(0))
+
+	// sender and message should match
+	sender, err := r.TestDAppV2ZEVM.SenderWithMessage(
+		&bind.CallOpts{},
+		[]byte(payloadOnRevert),
+	)
+	require.NoError(r, err)
+	require.Equal(r, r.ZEVMAuth.From, sender)
+
+	// the dApp address should get reverted amount
+	dAppBalanceAfter, err := r.SuiTokenZRC20.BalanceOf(&bind.CallOpts{}, dAppAddress)
+	require.NoError(r, err)
+	require.Equal(r, amount.Int64(), dAppBalanceAfter.Int64()-dAppBalanceBefore.Int64())
+}