+++ e2e/e2etests/test_sui_withdraw_and_call_revert_with_call.go
@@ -2,97 +2,80 @@ package e2etests
 
 import (
 	"math/big"
-	"time"
 
-	"cosmossdk.io/math"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/stretchr/testify/require"
 	"github.com/zeta-chain/protocol-contracts/pkg/gatewayzevm.sol"
 
 	"github.com/zeta-chain/node/e2e/runner"
 	"github.com/zeta-chain/node/e2e/utils"
+	"github.com/zeta-chain/node/testutil/sample"
 	crosschaintypes "github.com/zeta-chain/node/x/crosschain/types"
 )
 
-// TestSuiWithdrawRevertWithCall executes withdraw on zevm gateway.
-// The outbound is rejected by Sui network, and 'nonce_increase' is called instead to cancel the tx.
-func TestSuiWithdrawRevertWithCall(r *runner.E2ERunner, args []string) {
+// TestSuiWithdrawAndCallRevertWithCall executes withdrawAndCall on zevm gateway with SUI token.
+// The outbound is rejected by the connected module due to invalid payload (invalid address),
+// and 'onRevert' is called instead to handle the revert.
+func TestSuiWithdrawAndCallRevertWithCall(r *runner.E2ERunner, args []string) {
 	require.Len(r, args, 1)
-	amount := utils.ParseBigInt(r, args[0])
 
 	// ARRANGE
-	// given signer
-	signer, err := r.Account.SuiSigner()
-	require.NoError(r, err, "get deployer signer")
-	signerBalanceBefore := r.SuiGetSUIBalance(signer.Address())
+	// Given target package ID (example package) and a SUI amount
+	targetPackageID := r.SuiExample.PackageID.String()
+	amount := utils.ParseBigInt(r, args[0])
+
+	// create the payload for 'on_call' with invalid address
+	// taking the first 10 letters to form an invalid address
+	invalidAddress := sample.SuiAddress(r)[:10]
+	payloadOnCall, err := r.SuiCreateExampleWACPayload(invalidAddress)
+	require.NoError(r, err)
 
 	// given ZEVM revert address (the dApp)
 	dAppAddress := r.TestDAppV2ZEVMAddr
 	dAppBalanceBefore, err := r.SUIZRC20.BalanceOf(&bind.CallOpts{}, dAppAddress)
 	require.NoError(r, err)
 
-	// given random payload
-	payload := randomPayload(r)
-	r.AssertTestDAppEVMCalled(false, payload, amount)
-
-	// retrieve current zrc20 gas limit
-	oldGasLimit, err := r.SUIZRC20.GASLIMIT(&bind.CallOpts{})
-	require.NoError(r, err)
-	r.Logger.Info("current gas limit: %s", oldGasLimit.String())
-
-	// set a low ZRC20 gas limit so gasBudget will be low: "1000000"
-	// withdraw tx will be rejected due to execution error "InsufficientGas"
-	lowGasLimit := math.NewUintFromBigInt(big.NewInt(1000))
-	_, err = r.ZetaTxServer.UpdateZRC20GasLimit(r.SUIZRC20Addr, lowGasLimit)
-	require.NoError(r, err)
-
-	// wait for the new gas limit to take effect
-	utils.WaitForZetaBlocks(r.Ctx, r, r.ZEVMClient, 1, 10*time.Second)
+	// given random payload for 'onRevert'
+	payloadOnRevert := randomPayload(r)
+	r.AssertTestDAppEVMCalled(false, payloadOnRevert, amount)
 
 	// ACT
-	// approve the ZRC20
+	// approve SUI ZRC20 token
 	r.ApproveSUIZRC20(r.GatewayZEVMAddr)
 
-	// perform the withdraw with revert options
-	tx := r.SuiWithdrawSUI(
-		signer.Address(),
+	// perform the withdraw and call with revert options
+	tx := r.SuiWithdrawAndCallSUI(
+		targetPackageID,
 		amount,
+		payloadOnCall,
 		gatewayzevm.RevertOptions{
 			CallOnRevert:     true,
 			RevertAddress:    dAppAddress,
-			RevertMessage:    []byte(payload),
+			RevertMessage:    []byte(payloadOnRevert),
 			OnRevertGasLimit: big.NewInt(0),
 		},
 	)
-	r.Logger.EVMTransaction(*tx, "withdraw")
+	r.Logger.EVMTransaction(*tx, "withdraw_and_call")
 
 	// ASSERT
-	// wait for the CCTX to be mined
+	// wait for the cctx to be reverted
 	cctx := utils.WaitCctxMinedByInboundHash(r.Ctx, tx.Hash().Hex(), r.CctxClient, r.Logger, r.CctxTimeout)
+	r.Logger.CCTX(*cctx, "withdraw")
 	utils.RequireCCTXStatus(r, cctx, crosschaintypes.CctxStatus_Reverted)
 
 	// should have called 'onRevert'
-	r.AssertTestDAppZEVMCalled(true, payload, big.NewInt(0))
+	r.AssertTestDAppZEVMCalled(true, payloadOnRevert, big.NewInt(0))
 
 	// sender and message should match
 	sender, err := r.TestDAppV2ZEVM.SenderWithMessage(
 		&bind.CallOpts{},
-		[]byte(payload),
+		[]byte(payloadOnRevert),
 	)
 	require.NoError(r, err)
 	require.Equal(r, r.ZEVMAuth.From, sender)
-
-	// signer balance should remain unchanged in Sui chain
-	signerBalanceAfter := r.SuiGetSUIBalance(signer.Address())
-	require.Equal(r, signerBalanceBefore, signerBalanceAfter)
 
 	// the dApp address should get reverted amount
 	dAppBalanceAfter, err := r.SUIZRC20.BalanceOf(&bind.CallOpts{}, dAppAddress)
 	require.NoError(r, err)
 	require.Equal(r, amount.Int64(), dAppBalanceAfter.Int64()-dAppBalanceBefore.Int64())
-
-	// TEARDOWN
-	// restore old gas limit
-	_, err = r.ZetaTxServer.UpdateZRC20GasLimit(r.SUIZRC20Addr, math.NewUintFromBigInt(oldGasLimit))
-	require.NoError(r, err, "failed to restore gas limit")
 }